Fix another integer vs floating pt bug (this time in timestamps)

+++ RCCore/RCCorvisManagerFactory.mm
@@ -25,8 +25,8 @@ extern "C" {
 - (void)startPlugins;
 - (void)stopPlugins;
 - (void)receiveVideoFrame:(unsigned char*)pixel withWidth:(uint32_t)width withHeight:(uint32_t)height withTimestamp:(CMTime)timestamp;
-- (void)receiveAccelerometerData:(uint64_t)timestamp withX:(double)x withY:(double)y withZ:(double)z;
-- (void)receiveGyroData:(uint64_t)timestamp withX:(double)x withY:(double)y withZ:(double)z;
+- (void)receiveAccelerometerData:(double)timestamp withX:(double)x withY:(double)y withZ:(double)z;
+- (void)receiveGyroData:(double)timestamp withX:(double)x withY:(double)y withZ:(double)z;
 @end
 
 @implementation RCCorvisManagerImpl
@@ -138,7 +138,7 @@ extern "C" {
     }
 }
 
-- (void)receiveAccelerometerData:(uint64_t)timestamp withX:(double)x withY:(double)y withZ:(double)z
+- (void)receiveAccelerometerData:(double)timestamp withX:(double)x withY:(double)y withZ:(double)z
 {
     if (isPluginsStarted)
     {
@@ -152,7 +152,7 @@ extern "C" {
     }
 }
 
-- (void)receiveGyroData:(uint64_t)timestamp withX:(double)x withY:(double)y withZ:(double)z
+- (void)receiveGyroData:(double)timestamp withX:(double)x withY:(double)y withZ:(double)z
 {
     if (isPluginsStarted)
     {
